Created the namespaces mcrl2::core and mcrl2::core::detail for libraries defined in the core and core/detail respectively. Tools that use elements (functions, methods, data types, etc...) from these namespaces are adjusted.
Fixed unsigned vs. signed warnings in diagraphica.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4060 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/frame.cpp
@@ -2826,7 +2826,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlAttr->GetItemData( item ) != e.GetIndex() )
+                if ( (int) listCtrlAttr->GetItemData( item ) != e.GetIndex() )
                 {
                     listCtrlAttr->SetItemState(
                         item, 
@@ -2994,7 +2994,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlClust->GetItemData( item ) != e.GetIndex() )
+                if ( (int) listCtrlClust->GetItemData( item ) != e.GetIndex() )
                 {
                     listCtrlClust->SetItemState(
                         item, 
@@ -3023,7 +3023,7 @@ void Frame::onListCtrlSelect( wxListEvent &e )
                 wxLIST_STATE_SELECTED );
             while( item != -1 )
             {
-                if ( listCtrlClust->GetItemData( item ) != e.GetIndex() )
+                if ( (int) listCtrlClust->GetItemData( item ) != e.GetIndex() )
                 {
                     if ( selCount > 2 )
                     {
